Modified the <taggedRead> and <taggedWrite> to save the models in text mode. This makes the models portable across 32 and 64 bits OSs.
Of course, all the classes writing/reading models were changed for this in the library and in the examples.

	modified:   examples/data/models/facedetection/frontal/mct4-2-5-10-50-allface.cascade
	modified:   examples/data/models/facedetection/frontal/mct4-2-5-10-50.cascade
	modified:   examples/data/models/facedetection/frontal/mct4.cascade
	modified:   examples/data/models/facedetection/frontal/mct5-2-5-10-50-200.cascade
	modified:   examples/machines/README.txt
	modified:   examples/machines/convert3to5models.cc
	modified:   examples/proj/cb/examples.layout
	modified:   examples/scanning/testPatternList.cc
	modified:   src/boosting/IntLutMachine.cc
	modified:   src/boosting/RealLutMachine.cc
	modified:   src/boosting/StumpMachine.cc
	modified:   src/core/CmdLine.cc
	modified:   src/core/CmdOption.cc
	modified:   src/core/DataSet.h
	modified:   src/core/File.cc
	modified:   src/core/File.h
	modified:   src/core/FileListCmdOption.cc
	modified:   src/core/Machine.cc
	modified:   src/core/Tensor.h
	modified:   src/core/TensorFile.cc
	modified:   src/core/VariableCollector.cc
	modified:   src/ip/ipHLBPLienhart.cc
	modified:   src/ip/ipHaarLienhart.cc
	modified:   src/ip/ipLBP4R.cc
	modified:   src/ip/ipLBP8R.cc
	modified:   src/machine/CascadeMachine.cc
	modified:   src/machine/GradientMachine.cc
	modified:   src/machine/distributions/MultiVariateDiagonalGaussianDistribution.cc
	modified:   src/machine/distributions/MultiVariateMeansDistribution.cc
	modified:   src/machine/gradients/MLP.cc
	modified:   src/machine/normalisations/MeanVarNorm.cc
	modified:   src/proj/cb/torch5spro.cbp
	modified:   src/proj/cb/torch5spro.layout
	modified:   src/scanning/context/ContextDataSet.cc
	modified:   src/scanning/context/ContextMachine.cc
	modified:   src/scanning/context/LRMachine.cc

+++ src/core/Tensor.h
@@ -21,7 +21,7 @@ namespace Torch {
 			size[2] = 0;
 			size[3] = 0;
 		}
-		TensorSize(long dim0)
+		TensorSize(int dim0)
 		{
 			n_dimensions = 1;
 			size[0] = dim0;
@@ -30,7 +30,7 @@ namespace Torch {
 			size[3] = 0;
 		}
 
-		TensorSize(long dim0, long dim1)
+		TensorSize(int dim0, int dim1)
 		{
 			n_dimensions = 2;
 			size[0] = dim0;
@@ -39,7 +39,7 @@ namespace Torch {
 			size[3] = 0;
 		}
 
-		TensorSize(long dim0, long dim1, long dim2)
+		TensorSize(int dim0, int dim1, int dim2)
 		{
 			n_dimensions = 3;
 			size[0] = dim0;
@@ -48,7 +48,7 @@ namespace Torch {
 			size[3] = 0;
 		}
 
-		TensorSize(long dim0, long dim1, long dim2, long dim3)
+		TensorSize(int dim0, int dim1, int dim2, int dim3)
 		{
 			n_dimensions = 4;
 			size[0] = dim0;
@@ -58,7 +58,7 @@ namespace Torch {
 		}
 
 		int n_dimensions;
-		long size[4];
+		int size[4];
 	};
 
 	/////////////////////////////////////////////////////////////////
